requirements.txt
yfinance
pandas
requests
google-genai
beautifulsoup4
